+++ ready/example/lib/dashboard.dart
@@ -104,7 +104,7 @@ class DashBoardExample extends StatelessWidget {
               shimmer: false,
             );
           },
-          search: (value) {},
+          search: (context, v) {},
           icon: const Icon(Icons.list),
           id: 'list1',
           label: 'List',
@@ -113,7 +113,7 @@ class DashBoardExample extends StatelessWidget {
           builder: (Map<String, dynamic> parameters) {
             return const ReadyListExample(shimmer: true);
           },
-          search: (value) {},
+          search: (context, v) {},
           icon: const Icon(Icons.list),
           id: 'list2',
           label: 'List with shimmer',
@@ -126,7 +126,7 @@ class DashBoardExample extends StatelessWidget {
               builder: (Map<String, dynamic> parameters) {
                 return const ReadyGridExample(gridDelegate: Grids.columns_1);
               },
-              search: (value) {},
+              search: (context, v) {},
               icon: const Icon(Icons.local_cafe),
               id: 'grid 1',
               label: 'Grid 1',
@@ -135,7 +135,7 @@ class DashBoardExample extends StatelessWidget {
               builder: (Map<String, dynamic> parameters) {
                 return const ReadyGridExample(gridDelegate: Grids.columns_2);
               },
-              search: (value) {},
+              search: (context, v) {},
               icon: const Icon(Icons.local_cafe),
               id: 'grid 2',
               label: 'Grid 2',
@@ -144,7 +144,7 @@ class DashBoardExample extends StatelessWidget {
               builder: (Map<String, dynamic> parameters) {
                 return const ReadyGridExample(gridDelegate: Grids.columns_3);
               },
-              search: (value) {},
+              search: (context, v) {},
               icon: const Icon(Icons.local_cafe),
               id: 'grid 3',
               label: 'Grid 3',
@@ -153,7 +153,7 @@ class DashBoardExample extends StatelessWidget {
               builder: (Map<String, dynamic> parameters) {
                 return const ReadyGridExample(gridDelegate: Grids.columns_4);
               },
-              search: (value) {},
+              search: (context, v) {},
               icon: const Icon(Icons.local_cafe),
               id: 'grid 4',
               label: 'Grid 4',
@@ -162,7 +162,7 @@ class DashBoardExample extends StatelessWidget {
               builder: (Map<String, dynamic> parameters) {
                 return ReadyGridExample(gridDelegate: Grids.extent(150));
               },
-              search: (value) {},
+              search: (context, v) {},
               icon: const Icon(Icons.local_cafe),
               id: 'extent grid',
               label: 'Extent grid',
@@ -177,7 +177,7 @@ class DashBoardExample extends StatelessWidget {
               builder: (Map<String, dynamic> parameters) {
                 return ResponsiveList();
               },
-              search: (value) {},
+              search: (context, v) {},
               icon: const Icon(Icons.table_chart),
               id: 'responsive',
               label: 'Data table',
@@ -186,7 +186,7 @@ class DashBoardExample extends StatelessWidget {
               builder: (Map<String, dynamic> parameters) {
                 return const ReadyGridExample(gridDelegate: Grids.responsive);
               },
-              search: (value) {},
+              search: (context, v) {},
               icon: const Icon(Icons.local_cafe),
               id: 'grid responsive',
               label: 'Responsive grid',