+++ packages/ready_picker/CHANGELOG.md
@@ -13,3 +13,15 @@
 ## 1.5.3
 
 * Use ready version 1.5.3
+
+## 1.5.4
+
+* Use ready version 1.5.4
+
+## 1.5.5
+
+* Use ready version 1.5.5
+
+## 1.5.6
+
+* Use ready version 1.5.6